fix: remove unused class
